backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-community
langchain-core
